new NEW website going up!
so remember this website? Yeah, I didn't either. Well, over the past week or so, I took a renewed interest in it, so here's a major update. I remade the whole thing, using an actual UI framework. The old site was a pain to do anything with, style-wise, and the cross-browser compatibility (and mobile compatibility) wasn't good. But now I made this! Actually updated the content too, the old site was horribly out of date. Lots of new features, including a lightbox for images, proper responsiveness, a really cool parallax animation on the homepage header, and the good news is, the site should still be quite fast! Enjoy :)